+++ rec/prepare-fzf.sh
@@ -10,5 +10,6 @@ source_dir=$1
 output=$2
 output_dir=${output%/*}
 mkdir -p "$output_dir"
-cargo build --quiet --release --example fzf --manifest-path "$source_dir/Cargo.toml"
+cargo build --locked --quiet --release --example fzf --features tracing \
+    --manifest-path "$source_dir/Cargo.toml"
 cp "$source_dir/target/release/examples/fzf" "$output"